Implement ProposerSlashingValidator for networking layer

+++ ethereum/datastructures/src/testFixtures/java/tech/pegasys/teku/datastructures/util/DataStructureUtil.java
@@ -96,7 +96,7 @@ public final class DataStructureUtil {
   }
 
   public long randomLong() {
-    return new Random(nextSeed()).nextLong();
+    return new Random(nextSeed()).nextInt();
   }
 
   public UnsignedLong randomUnsignedLong() {
